Implemented output logging for passing steps (failing steps pending).
