+++ source/Xunit.Gherkin.Quick/CoreModel/Scenario/IScenarioOutput.cs
