+++ source/Xunit.Gherkin.Quick/CoreModel/Scenario/ScenarioOutput.cs
@@ -1,0 +1,19 @@
+﻿using Xunit.Abstractions;
+
+namespace Xunit.Gherkin.Quick
+{
+    internal sealed class ScenarioOutput : IScenarioOutput
+    {
+        private readonly ITestOutputHelper _testOutputHelper;
+
+        public ScenarioOutput(ITestOutputHelper testOutputHelper)
+        {
+            _testOutputHelper = testOutputHelper ?? throw new System.ArgumentNullException(nameof(testOutputHelper));
+        }
+
+        public void StepPassed(string stepText)
+        {
+            _testOutputHelper.WriteLine($"{stepText}: PASSED");
+        }
+    }
+}
