+++ source/Xunit.Gherkin.Quick/CoreModel/ScenarioExecutor.cs
@@ -23,7 +23,7 @@ namespace Xunit.Gherkin.Quick
             var featureFile = _featureFileRepository.GetByFilePath(featureClass.FeatureFilePath);
 
             var scenario = featureClass.ExtractScenario(scenarioName, featureFile);
-            await scenario.ExecuteAsync(null);
+            await scenario.ExecuteAsync(new ScenarioOutput(featureInstance.Output));
         }
     }
 }
